[Refactor] Split expression extensions from core Expr module
