+++ be/src/connector/mysql/AGENTS.md
@@ -9,7 +9,7 @@ Run `python3 build-support/check_be_module_boundaries.py --mode full` to validat
 MySQL connector implementation above connector contracts without registry composition, storage, service, or full Exec coupling.
 - Targets: `ConnectorMySQL`
 - Allowed internal include prefixes: `connector/mysql/`, `connector/connector.h`, `connector/data_source.h`, `connector/data_source_provider.h`, `exprs/`, `runtime/`, `column/`, `types/`, `common/`, `base/`, `gutil/`, `gen_cpp/`
-- Allowed target deps: `ConnectorPrimitive`, `ExprCore`, `Runtime`, `ChunkCore`, `ColumnCore`, `Types`, `Common`, `Base`, `Gutil`, `StarRocksGen`
+- Allowed target deps: `ConnectorPrimitive`, `Expr`, `Runtime`, `ChunkCore`, `ColumnCore`, `Types`, `Common`, `Base`, `Gutil`, `StarRocksGen`
 - Core tests: `connector_mysql_test`
 - Remediation: Keep ConnectorMySQL limited to MySQL read-side scan and scanner logic; move registration into ConnectorBootstrap and avoid pulling Connector, storage, service, or full Exec code into the connector library.
 <!-- END GENERATED: BE MODULE HARNESSES -->
