+++ be/src/exprs_ext/table_function/list_rowsets.cpp
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "exprs/table_function/list_rowsets.h"
+#include "exprs_ext/table_function/list_rowsets.h"
 
 #include <algorithm>
