+++ be/src/exprs_ext/utility/utility_functions.cpp
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "exprs/utility_functions.h"
+#include "exprs_ext/utility/utility_functions.h"
 
 #include "column/column_visitor_adapter.h"
 #include "common/util/thrift_client_cache.h"
