+++ be/test/compute_env/CMakeLists.txt
@@ -95,7 +95,7 @@ set(COMPUTE_ENV_SORTING_TEST_FILES
 
 set(COMPUTE_ENV_SORTING_TEST_LINK_LIBS
     ComputeEnv
-    ExprCore
+    Expr
     Runtime
     ChunkCore
     ColumnCore
