+++ be/test/exprs/dict_expr_test.cpp
@@ -28,9 +28,9 @@
 #include "common/object_pool.h"
 #include "compute_env/global_dict/fragment_dict_state.h"
 #include "exprs/column_ref.h"
-#include "exprs/dictmapping_expr.h"
 #include "exprs/expr_context.h"
 #include "exprs/placeholder_ref.h"
+#include "exprs_ext/dict/dictmapping_expr.h"
 #include "gen_cpp/Types_types.h"
 #include "runtime/exception.h"
 #include "runtime/mem_pool.h"
