+++ be/test/exprs/exprs_test_helper.h
@@ -20,7 +20,7 @@
 #include "column/chunk.h"
 #include "exprs/array_expr.h"
 #include "exprs/expr_executor.h"
-#ifndef STARROCKS_EXPR_CORE_TEST_NO_JIT
+#ifndef STARROCKS_EXPR_TEST_NO_JIT
 #include "exprs/jit/jit_expr.h"
 #endif
 #include "gen_cpp/Descriptors_types.h"
@@ -160,7 +160,7 @@ public:
         // Verify the original result.
         test_func(ptr);
 
-#ifdef STARROCKS_EXPR_CORE_TEST_NO_JIT
+#ifdef STARROCKS_EXPR_TEST_NO_JIT
         (void)expr;
         (void)runtime_state;
         (void)need_jit;
@@ -194,7 +194,7 @@ public:
     }
 
     static bool should_verify_with_jit(const Expr* expr, RuntimeState* runtime_state) {
-#ifdef STARROCKS_EXPR_CORE_TEST_NO_JIT
+#ifdef STARROCKS_EXPR_TEST_NO_JIT
         (void)expr;
         (void)runtime_state;
         return false;
@@ -204,7 +204,7 @@ public:
     }
 
     static void verify_result_with_jit(const ColumnPtr& ptr, Expr* expr, RuntimeState* runtime_state) {
-#ifdef STARROCKS_EXPR_CORE_TEST_NO_JIT
+#ifdef STARROCKS_EXPR_TEST_NO_JIT
         (void)ptr;
         (void)expr;
         (void)runtime_state;
