+++ be/test/formats/CMakeLists.txt
@@ -71,7 +71,7 @@ set(FORMAT_TEST_LINK_LIBS
     FormatJson
     FormatCsv
     FormatCore
-    ExprCore
+    Expr
     Runtime
     FileSystem
     ChunkCore
