+++ be/test/storage/dictionary_cache_manager_test.cpp
@@ -25,8 +25,8 @@
 #include "compute_env/compute_env.h"
 #include "compute_env/dictionary_cache/chunk_util.h"
 #include "exec/exec_env.h"
-#include "exprs/dictionary_get_expr.h"
 #include "exprs/mock_vectorized_expr.h"
+#include "exprs_ext/dict/dictionary_get_expr.h"
 #include "runtime/descriptor_helper.h"
 #include "runtime/runtime_state.h"
 #include "storage/chunk_helper.h"
