+++ be/test/storage_primitive/CMakeLists.txt
@@ -41,7 +41,7 @@ set(STORAGE_PRIMITIVE_TEST_FILES
 
 set(STORAGE_PRIMITIVE_TEST_LINK_LIBS
     StoragePrimitive
-    ExprCore
+    Expr
     Runtime
     ChunkCore
     ColumnCore
